Added 'pattern' test to w10n tests
The test coads_SST_data.bescmd results in an (intentional) error. I
added a Pattern rule to the *.at test file so that the build options,
etc., will not make differences in the output that will make the test
erroneously fail. I only modified one test to use the new rule; all
tests pass on my version of the master branch.